feat: 입력 숫자 중복 확인 테스트 코드 추가
